+++ FontVaporArq/Models/Resultado.swift
@@ -7,6 +7,6 @@
 import Foundation
 
 enum Resultado {
-    case sucesso(novoValor: Decimal)
+    case sucesso(novoValor: Double)
     case falha(erro: String)
 }
